spaces_cache/tasal9_ZamAI-Phi3-Mini-Pashto-Demo/requirements.txt
httpx
beautifulsoup4
pypdf
transformers
accelerate
peft
datasets
sentence-transformers
faiss-cpu
gradio
numpy
tqdm
spaces
torch>=2.1.0
